Reduced channeler's interface + Funder::ephemeral is now private.
