+++ components/channeler/Cargo.toml
@@ -12,6 +12,7 @@ crypto = { path = "../crypto", version = "0.1.0", package = "offst-crypto" }
 identity = { path = "../identity", version = "0.1.0" , package = "offst-identity" }
 timer = { path = "../timer", version = "0.1.0" , package = "offst-timer" }
 proto = { path = "../proto", version = "0.1.0" , package = "offst-proto" }
+relay = { path = "../relay", version = "0.1.0" , package = "offst-relay" }
 
 log = "0.4"
 futures-preview = "0.3.0-alpha.11"
