+++ components/funder/src/handler/tests/change_address.rs
@@ -39,9 +39,9 @@ async fn task_handler_change_address(identity_client1: IdentityClient,
     };
 
     let mut state1 = FunderState::<u32>::new(&pk1, &0x1337u32);
-    let mut ephemeral1 = Ephemeral::new(&state1);
+    let mut ephemeral1 = Ephemeral::new();
     let mut state2 = FunderState::<u32>::new(&pk2, &0x1338u32);
-    let mut ephemeral2 = Ephemeral::new(&state2);
+    let mut ephemeral2 = Ephemeral::new();
 
     let mut rng = RngContainer::new(DummyRandom::new(&[3u8]));
