+++ components/funder/src/handler/tests/pair_basic.rs
@@ -41,9 +41,9 @@ async fn task_handler_pair_basic<'a>(identity_client1: &'a mut IdentityClient,
     };
 
     let mut state1 = FunderState::<u32>::new(&pk1, &11u32);
-    let mut ephemeral1 = Ephemeral::new(&state1);
+    let mut ephemeral1 = Ephemeral::new();
     let mut state2 = FunderState::<u32>::new(&pk2, &22u32);
-    let mut ephemeral2 = Ephemeral::new(&state2);
+    let mut ephemeral2 = Ephemeral::new();
 
     let mut rng = RngContainer::new(DummyRandom::new(&[3u8]));
