+++ components/funder/src/handler/tests/pair_inconsistency.rs
@@ -40,9 +40,9 @@ async fn task_handler_pair_inconsistency<'a>(identity_client1: &'a mut IdentityC
     };
 
     let mut state1 = FunderState::<u32>::new(&pk1, &0x1337u32);
-    let mut ephemeral1 = Ephemeral::new(&state1);
+    let mut ephemeral1 = Ephemeral::new();
     let mut state2 = FunderState::<u32>::new(&pk2, &0x1338u32);
-    let mut ephemeral2 = Ephemeral::new(&state2);
+    let mut ephemeral2 = Ephemeral::new();
 
     let mut rng = RngContainer::new(DummyRandom::new(&[3u8]));
