+++ components/node/src/adapters.rs
@@ -48,75 +48,6 @@ where
     }
 }
 
-/// A connection style encrypt transform.
-/// Does not return the public key of the remote side, because we already know it.
-#[derive(Clone)]
-pub struct ConnectEncryptTransform<ET> {
-    encrypt_transform: ET,
-}
-
-impl<ET> ConnectEncryptTransform<ET> {
-    pub fn new(encrypt_transform: ET) -> Self {
-
-        ConnectEncryptTransform {
-            encrypt_transform,
-        }
-    }
-}
-
-impl<ET> FutTransform for ConnectEncryptTransform<ET> 
-where
-    ET: FutTransform<Input=(Option<PublicKey>, ConnPairVec),Output=Option<(PublicKey, ConnPairVec)>> + Send,
-{
-    type Input = (PublicKey, ConnPairVec);
-    type Output = Option<ConnPairVec>;
-
-    fn transform(&mut self, input: Self::Input)
-        -> BoxFuture<'_, Self::Output> {
-
-        let (public_key, conn_pair) = input;
-
-        Box::pin(async move {
-            let (_public_key, conn_pair) = await!(
-                self.encrypt_transform.transform((Some(public_key), conn_pair)))?;
-            Some(conn_pair)
-        })
-    }
-}
-
-/// A Listen style encrypt transform.
-/// Returns the public key of the remote side, because we can not predict it.
-#[derive(Clone)]
-pub struct ListenEncryptTransform<ET> {
-    encrypt_transform: ET,
-}
-
-impl<ET> ListenEncryptTransform<ET> {
-    pub fn new(encrypt_transform: ET) -> Self {
-
-        ListenEncryptTransform {
-            encrypt_transform,
-        }
-    }
-}
-
-impl<ET> FutTransform for ListenEncryptTransform<ET> 
-where
-    ET: FutTransform<Input=(Option<PublicKey>, ConnPairVec),Output=Option<(PublicKey, ConnPairVec)>> + Send,
-{
-    type Input = (PublicKey, ConnPairVec);
-    type Output = Option<(PublicKey, ConnPairVec)>;
-
-    fn transform(&mut self, input: Self::Input)
-        -> BoxFuture<'_, Self::Output> {
-
-        let (public_key, conn_pair) = input;
-
-        Box::pin(async move {
-            await!(self.encrypt_transform.transform((Some(public_key), conn_pair)))
-        })
-    }
-}
 
 
 // C: FutTransform<Input=IndexServerAddress, Output=Option<ServerConn>> + Send,
